### 2021-01-08 #2947 @whedon
[PRE REVIEW]: Coral: a parallel spectral solver for fluid dynamics and partial differential equations
**Submitting author:** @benmql (<a href="http://orcid.org/0000-0001-6283-0382">BENJAMIN MIQUEL</a>)
**Repository:** <a href="https://github.com/BenMql/coral" target ="_blank">https://github.com/BenMql/coral</a>
**Version:** v1.0.1
