+++ etc/python_scripts/read_timeseries.py
@@ -5,9 +5,9 @@ import matplotlib.pyplot as plt
 # Read timeseries
 #..........................................................................
 time = np.fromfile('Timeseries/time.dat',  dtype=np.float_)
-uu = np.fromfile('Timeseries/uu_full.dat', dtype=np.float_)
-vv = np.fromfile('Timeseries/vv_full.dat', dtype=np.float_)
-ww = np.fromfile('Timeseries/ww_full.dat', dtype=np.float_)
+uu = np.fromfile('Timeseries/uu_volAvg.dat', dtype=np.float_)
+vv = np.fromfile('Timeseries/vv_volAvg.dat', dtype=np.float_)
+ww = np.fromfile('Timeseries/ww_volAvg.dat', dtype=np.float_)
 
 # =========================================================================
 # check that all data has same size 
