### 2021-01-14 e469b83
adding python script for slices
