### 2021-01-19 #2947 @BenMql
@eloisabentivegna Thank you for accepting to handle the submission as an editor.

As suggested, I include below a few suggestions for referees (selected in the list of JOSS referees):
- mikaem
- harpolea
- p-costa
In addition to existing referees, the following scientists would be suitable for reviewing:
- Dr. Adrian Barker (U. Leeds), https://eps.leeds.ac.uk/maths/staff/4006/dr-adrian-j-barker
- Dr. Celine Guervilly (U. Newcastle), https://www.ncl.ac.uk/maths-physics/staff/profile/celineguervilly.html#background
- Dr. Thomas Gastine (CNRS-IPGP), http://www.ipgp.fr/~gastine/
- Dr. Benjamin Favier (Aix-Marseille U., CNRS), https://sites.google.com/site/bfavierhome/home?authuser=0

Thank you! --BM
