+++ README.md
@@ -1,12 +1,12 @@
 ![baniere.png](http://www.normalesup.org/~benmiquel/misc/baniere.png)
 
-# Coral v1.0
+# Coral v1.1
 
 
 Coral is a spectral PDEs solver for the plane-layer geometry, written in modern Fortran and highly scalable. 
 The name comes from the initial motivation, which was **Co**nvection in **Ra**pidly rotating **L**ayers (eventhough the code has now a much more general scope).
 
-The current version is v1.0.1 
+The current version is v1.1.0
 
 
 ## Tutorials
@@ -16,7 +16,8 @@ The current version is v1.0.1
 + [Data vizualisation](https://github.com/BenMql/coral/wiki/4_Data_visualisation)
 
 ## Reference
-+ [Input files](https://github.com/BenMql/coral/wiki/5_Input_files)
++ [Input files--General overview](https://github.com/BenMql/coral/wiki/5_Input_files). 
++ [Input files--Equations](https://github.com/BenMql/coral/tree/master/doc): python notebooks with TeX formulas
 + [Common issues](https://github.com/BenMql/coral/wiki/6_Common_issues)
 
 ## Benchmarks
