### 2021-01-22 c3f09d6
fixing minor bugs in .py scripts
