### 2021-02-02 BenMql/coral#1 @robertsawko
Include DOI for the APS paper
Hi Ben,

I started reviewing [your JOSS submission](https://github.com/openjournals/joss-reviews/issues/2978) today. Very interesting piece of research and congratulations on your results and publications so far.

Minor correction. As per bot suggestion can you add the DOI to the references for your APS paper:

```
10.1103/physrevfluids.4.121501
```
