### 2021-02-04 3880c11
issue #6--fixing makefile
