### 2021-02-10 43236c6
minor timings fix (smaller dt when restarting)
