### 2021-02-10 e8d7b74
colorplot_slice.py is now python-version agnostic (works with both python2 and python3)
