+++ etc/benchmarks/02_RRBC_julien_JFM96/compute_nusselt.py
@@ -1,0 +1,82 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+# =========================================================================
+# Read timeseries
+#..........................................................................
+time = np.fromfile('Timeseries/time.dat',  dtype=np.float_)
+Ttop = np.fromfile('Timeseries/tFull_XYavg_z00001.dat', dtype=np.float_)[:-2]
+Tbot = np.fromfile('Timeseries/tFull_XYavg_z00072.dat', dtype=np.float_)[:-2]
+
+quarterLength = time.shape[0]//4
+
+# =========================================================================
+# compute dz
+#..........................................................................
+dz = 1.-0.5 - 0.5*np.cos(np.pi/2/72.)
+
+nusseltTop_1stQuarter = np.sum(np.diff(time)[0*quarterLength:1*quarterLength]*
+                                        Ttop[0*quarterLength:1*quarterLength]
+                              ) / np.sum(np.diff(time)[0*quarterLength:1*quarterLength])/dz
+nusseltBot_1stQuarter =-np.sum(np.diff(time)[0*quarterLength:1*quarterLength]*
+                                        Tbot[0*quarterLength:1*quarterLength]
+                              ) / np.sum(np.diff(time)[0*quarterLength:1*quarterLength])/dz
+nusseltTop_2ndQuarter = np.sum(np.diff(time)[1*quarterLength:2*quarterLength]*
+                                        Ttop[1*quarterLength:2*quarterLength]
+                              ) / np.sum(np.diff(time)[1*quarterLength:2*quarterLength])/dz
+nusseltBot_2ndQuarter =-np.sum(np.diff(time)[1*quarterLength:2*quarterLength]*
+                                        Tbot[1*quarterLength:2*quarterLength]
+                              ) / np.sum(np.diff(time)[1*quarterLength:2*quarterLength])/dz
+nusseltTop_3rdQuarter = np.sum(np.diff(time)[2*quarterLength:3*quarterLength]*
+                                        Ttop[2*quarterLength:3*quarterLength]
+                              ) / np.sum(np.diff(time)[2*quarterLength:3*quarterLength])/dz
+nusseltBot_3rdQuarter =-np.sum(np.diff(time)[2*quarterLength:3*quarterLength]*
+                                        Tbot[2*quarterLength:3*quarterLength]
+                              ) / np.sum(np.diff(time)[2*quarterLength:3*quarterLength])/dz
+nusseltTop_4thQuarter = np.sum(np.diff(time)[3*quarterLength:4*quarterLength]*
+                                        Ttop[3*quarterLength:4*quarterLength]
+                              ) / np.sum(np.diff(time)[3*quarterLength:4*quarterLength])/dz
+nusseltBot_4thQuarter =-np.sum(np.diff(time)[3*quarterLength:4*quarterLength]*
+                                        Tbot[3*quarterLength:4*quarterLength]
+                              ) / np.sum(np.diff(time)[3*quarterLength:4*quarterLength])/dz
+
+
+mean_nusselt = (nusseltTop_2ndQuarter+
+                nusseltBot_2ndQuarter+
+                nusseltTop_3rdQuarter+
+                nusseltBot_3rdQuarter+
+                nusseltTop_4thQuarter+
+                nusseltBot_4thQuarter)
+mean_nusselt /= 6.
+std_nusselt = np.sqrt(1./6.*(
+                (nusseltTop_2ndQuarter-mean_nusselt)**2+
+                (nusseltBot_2ndQuarter-mean_nusselt)**2+
[…]
