+++ etc/benchmarks/02_Taylor_Couette/coral.parameters.in
@@ -1,19 +1,0 @@
-0           ! ........... RESOLUTION ............ !
-96          ! 01 ! (INT) ... along X         ! 01 !
-144         ! 02 ! (INT) ... along Y         ! 02 !
-32          ! 03 ! (INT) ... along Z         ! 03 !
-0           ! ............ GEOMETRY ............. !
-4.d0 ! 05 ! (DBL) x/z aspect ratio    ! 05 !
-6.283185307179586      ! 06 ! (DBL) y/z aspect ratio    ! 06 !
-0.485d0        ! 07 ! (DBL) gap (should=1.0_dp) ! 07 !
-3.7925d0       ! 08 ! (DBL) center              ! 08 !
-0           ! .......... TIMESTEPPING ........... !
-1.d0    ! 10 ! (DBL) maximum timestep    ! 10 !
-0.3d0       ! 11 ! (DBL) timestep tolerance  ! 11 !
-0.4d0       ! 12 ! (DBL) CFL Factor          ! 12 !
-'ARS_222'   ! 13 ! scheme                    ! 13 !
-0           ! ......... QUICKSAVE RESTART ....... !
-.true.      ! 15 ! (BOO) quicksave restart   ! 15 !
-0           ! ............ TIMER ................ !
-7200.d0   ! 17 ! (DBL) Time limit (in sec.)! 17 !
-
