### 2021-02-23 #2978 @rhaas80
comments working through installation:
* in the git clone command `git clone https://github.com/BenMql/coral.git .` I would either leave out the dot `.` or replace by `$PWD` which is easier to see
* `export CORAL_ROOT=$(PWD)` is incorrect and must be `$PWD` or `${PWD}` or `$(pwd)` since there is no `PWD` executable only, a `PWD` variable and a `pwd` executable.
* no link is provided for downloading MKL, would benice to have though Intel seems to restructure their website periodically making this harder
* MKL is not open-source (it is free-as-in-beer but not free-as-in-speech), would be nice to have alternative, also for non-Intel CPUs
* Decomp2d seems to have its own set of dependencies eg FFTE that the `Decomp2d_installation` file does not mention
* Decomp2d's makefile refers to `$(MKL_PATH)/lib/ia32` though by now `intel64` ought to be the default choice
* unclear if Decomp2d must be built against FFTW3 (coral is) or if using MKL's FFTW wrapper are ok
* FAILURE: at least the `clean` target assumes that `CORAL_ROOT` ends in a slash, which is not the case for `export CORAL_ROOT=$PWD`
* FAILURE: gfortran-10 fails with:
```
/src/sparse_tools/sparse_manipulations.f90:124:19:

  124 |                    ptr2B, A%ncol, dbeta, ptr2C, A%nrow)
      |                   1
......
  189 |                    ptr2B, A%ncol, dbeta, ptr2C, A%nrow)
      |                   2
Error: Type mismatch between actual argument at (1) and actual argument at (2) (COMPLEX(8)/REAL(8)).
```
* needs either fix in code or `-fallow-argument-mismatch` in `MPIFLAGS` (proper code being preferred)
* OPTIONAL: parallel build using `make -j` fails since dependencies for module files are not recorded in `makefile`
