### 2021-02-23 #2978 @BenMql
Thank you for your comments.
+ `.`, `$PWD` and missing slashes issues have been fixed.
+ the `2decomp` doc now states explicitly that a vanilla fftw is required.
+ A link to MKL has been added.
+ I share your enthusiasm for "BLAS/LAPACK-vendor" agnostic code. Down the road, my hope is that Coral can be linked to any BLAS/LAPACK library. For the time being though, I stick to `MKL` as some of their (non standard) routines are used.
+ working on the failures. This is interesting, I never detected them despite an heavy use of `pedantic` flags and standard-compliance checks (not with gfortran-10 though).
