### 2021-02-23 e24163f
fixed type in sparse_manipulations.f90
