### 2021-02-24 #2978 @rhaas80
The argument mismatch is new by gfortran. Apparently such usage was never really allowed by the Fortran standard, but also never really enforced by compilers (though you may want to give eg the Cray compiler a try which in my experience is pickier than gfortran or ifort). You are certainly not the only code tripped up by this, eg https://www.google.com/search?q=gfortran-10+allow-mismatch&ie=utf-8&oe=utf-8 . For gfortran's release announcement about the issue, please see: https://gcc.gnu.org/gcc-10/changes.html#fortran
