### 2021-02-24 #2978 @BenMql
Thank you for these details. This feature definitely helped catching a bug which fortunately had no consequences. (Some pointers used to be twice as long as intended, but fortunately their second half was never accessed.)
Anyway, this glitch is fixed now. Let me know if you run into more trouble while compiling.
