+++ paper/paper.bib
@@ -19,7 +19,9 @@ pages={R1}}
   pages = {121501},
   numpages = {9},
   year = {2019},
-  month = {Dec}
+  month = {Dec},
+  publisher = {American Physical Society},
+  DOI = {10.1103/physrevfluids.4.121501}
 }
 
 
@@ -31,3 +33,122 @@ pages={R1}}
   journal = {Cray user group 2010 conference},
   url = {http://www.2decomp.org/pdf/17B-CUG2010-paper-Ning_Li.pdf}
 }
+
+@book{chandra,
+  author = {Chandrasekhar, S.},
+  title = {Hydrodynamic and Hydromagnetic Stability},
+  publisher = {Clarendon Press: Oxford University Press},
+  year = {1961}
+  }
+
+@article{julieniJFM96,
+  title={Rapidly rotating turbulent Rayleigh-B{\'e}nard convection},
+  author={Julien, K. and Legg, S. and McWilliams, J. and Werne, J.},
+  journal={J. Fluid Mech.},
+  volume={322},
+  pages={243--273},
+  year={1996},
+  publisher={Cambridge University Press},
+  doi = {10.1017/S0022112096002789}
+}
+
+@article{cooperPRF20,
+  title = {Subcritical dynamos in rapidly rotating planar convection},
+  author = {Cooper, R. G. and Bushby, P. J. and Guervilly, C.},
+  journal = {Phys. Rev. Fluids},
+  volume = {5},
+  issue = {11},
+  pages = {113702},
+  numpages = {23},
+  year = {2020},
+  month = {Nov},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevFluids.5.113702}
+}
+
+@article{dedalus,
+  title = {Dedalus: A flexible framework for numerical simulations with spectral methods},
+  author = {Burns, Keaton J. and Vasil, Geoffrey M. and Oishi, Jeffrey S. and Lecoanet, Daniel and Brown, Benjamin P.},
+  journal = {Phys. Rev. Research},
+  volume = {2},
+  issue = {2},
+  pages = {023068},
+  numpages = {39},
+  year = {2020},
+  month = {Apr},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevResearch.2.023068}
+}
+
+@article{nektar++,          
+title = {Nektar++: Enhancing the capability and application of high-fidelity spectral/hp element methods},
+journal = {Computer Physics Communications},
+volume = {249},
+pages = {107110},
+year = {2020},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2019.107110},
+author = {David Moxey and Chris D. Cantwell and Yan Bao and Andrea Cassinelli and Giacomo Castiglioni and Sehun Chun and Emilia Juda and Ehsan Kazemi and Kilian Lackhove and Julian Marcon and Gianmarco Mengaldo and Douglas Serson and Michael Turner and Hui Xu and Joaquim Peiró and Robert M. Kirby and Spencer J. Sherwin}
+}
+
+@article{nek5000,
+author={Fisher, P. F., Lottes, J. W., Kerkemeier, S. G.},
+url={http://nek5000.mcs.anl.gov}}
+
+@article{spectralDNS,
+  doi = {10.21105/joss.01071},
+  url = {https://doi.org/10.21105/joss.01071},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {31},
+  pages = {1071},
+  author = {Mikael Mortensen},
+  title = {Shenfun: High performance spectral Galerkin computing platform},
+  journal = {Journal of Open Source Software}
+}
+
+@article{freefem++,
+        AUTHOR = {Hecht, F.},
[…]
