+++ paper/paper.bib
@@ -41,7 +41,7 @@ pages={R1}}
   year = {1961}
   }
 
-@article{julieniJFM96,
+@article{julienJFM96,
   title={Rapidly rotating turbulent Rayleigh-B{\'e}nard convection},
   author={Julien, K. and Legg, S. and McWilliams, J. and Werne, J.},
   journal={J. Fluid Mech.},
@@ -80,7 +80,7 @@ pages={R1}}
   doi = {10.1103/PhysRevResearch.2.023068}
 }
 
-@article{nektar++,          
+@article{nektar,          
 title = {Nektar++: Enhancing the capability and application of high-fidelity spectral/hp element methods},
 journal = {Computer Physics Communications},
 volume = {249},
@@ -108,13 +108,13 @@ url={http://nek5000.mcs.anl.gov}}
   journal = {Journal of Open Source Software}
 }
 
-@article{freefem++,
-        AUTHOR = {Hecht, F.},
-        TITLE = {New development in FreeFem++},
-        JOURNAL = {J. Numer. Math.},
-        FJOURNAL = {Journal of Numerical Mathematics},
-        VOLUME = {20}, YEAR = {2012},
-        NUMBER = {3-4}, PAGES = {251--265}
+@article{freefem,
+        author = {Hecht, F.},
+        title = {New development in FreeFem++},
+        journal = {J. Numer. Math.},
+        volume = {20}, year = {2012},
+        number = {3-4}, pages = {251--265},
+	doi={10.1515/jnum-2012-0013}
 }
 
 
@@ -124,8 +124,7 @@ url={http://nek5000.mcs.anl.gov}}
         year = {2015},
         journal = {Archive of Numerical Software},
         volume = {3},
-        number = {100},
-        page = {9-23}
+        number = {100}
 }
 
 @article{fluiddyn,
