### 2021-03-04 #2978 @rhaas80
In "2_First_run":
* typo" "After edition" should be "After editing"
* should indicate how much memory the sample run takes
* trying the test data on any number of MPI ranks > 1 I get a SEGFAULT with a backtrace and some information that seems to indicate that `opt_decomp` passed to `transpose_y_to_z` is incorect. Here's the information:
```
Thread 1 "coral_LP.exe" received signal SIGSEGV, Segmentation fault.
__memmove_avx_unaligned_erms () at ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S:262
262     ../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S: No such file or directory.
(gdb) up
#1  0x00007fffecd982fc in ?? () from /usr/lib/x86_64-linux-gnu/libopen-pal.so.40
(gdb) up
#2  0x00007fffecf268aa in ompi_datatype_sndrcv () from /usr/lib/x86_64-linux-gnu/libmpi.so.40
(gdb)
#3  0x00007fffecf732e3 in ompi_coll_base_alltoallv_intra_pairwise ()
   from /usr/lib/x86_64-linux-gnu/libmpi.so.40
(gdb)
#4  0x00007fffe65b0fa2 in ompi_coll_tuned_alltoallv_intra_dec_fixed ()
   from /usr/lib/x86_64-linux-gnu/openmpi/lib/openmpi3/mca_coll_tuned.so
(gdb)
#5  0x00007fffecf2936b in PMPI_Alltoallv () from /usr/lib/x86_64-linux-gnu/libmpi.so.40
(gdb)
#6  0x00007fffed034d0b in pmpi_alltoallv__ () from /usr/lib/x86_64-linux-gnu/libmpi_mpifh.so.40
(gdb)
#7  0x00005555555de192 in decomp_2d::transpose_y_to_z_real (
    src=<error reading variable: value requires 2654208 bytes, which is more than max-value-size>,
    dst=<error reading variable: value requires 2654208 bytes, which is more than max-value-size>,
    opt_decomp=...) at transpose_y_to_z.f90:87
87               real_type, DECOMP_2D_COMM_ROW, ierror)
(gdb) print opt_decomp
$1 = ( xst = (1, 1, 49), xen = (72, 96, 96), xsz = (72, 96, 48), yst = (1, 1, 49), yen = (72, 96, 96), ysz = (72, 96, 48), zst = (1, 49, 1), zen = (72, 96, 96), zsz = (72, 48, 96), x1dist = (72), y1dist = (96), y2dist = (48, 48), z2dist = (48, 48), x1cnts = (331776), y1cnts = (331776), y2cnts = (165888, 165888), z2cnts = (165888, 165888), x1disp = (0), y1disp = (0), y2disp = (0, 165888), z2disp = (0, -322823200), x1count = 331776, y1count = 331776, y2count = 165888, z2count = 165888, even = .TRUE. )
```
where `opt_decomp%z2disp(2)` being `-322823200` seems like the likely culprit to me.
