### 2021-03-04 #2978 @rhaas80
This seems to be due to assuming that `allocate` would zero initialize the allocated memory. I can make the SEGFAULT go away by adding
```
    decomp%x1cnts = 0
    decomp%y1cnts = 0
    decomp%y2cnts = 0
    decomp%z2cnts = 0
    decomp%x1disp = 0
    decomp%y1disp = 0
    decomp%y2disp = 0
    decomp%z2disp = 0
```
just before the `prepare_buffer` call.
