### 2021-03-04 #2978 @BenMql
Roland,

I have fixed the wiki `2_first_run`.

I am most grateful for your checks, comments and suggestions.
In `2decomp&fft`, the arrays of integers are now initialized to 0 as you suggested (available as the most recent commit to `https://github.com/benmql/2decomp_emptyFourierFourier'). I will run the code through valgrind as you suggest, probably some time next week.
