### 2021-03-07 #2978 @eloisabentivegna
Thanks @rhaas80 for your thorough comments, and @BenMql for addressing them promptly!

@robertsawko, do you have any update to report at this stage?
