### 2021-03-15 #2978 @rhaas80
> Thanks @rhaas80 for your thorough comments, and @BenMql for addressing them promptly!
@eloisabentivegna  Sorry about the long delay. I am unfortunately not yet done with the review.
