### 2021-03-16 #2978 @robertsawko
Apologies for a radio silence. I am going to get back to reviewing this week. I have been reporting on GH issues of the software repository as I think that's the recommendation with JOSS. Made a few comments about the documentation. Unfortunately I run into an issue with the build on Arch Linux, but I think I should be able to resolve it now.
