### 2021-03-16 #2978 @robertsawko
Hello @eloisabentivegna,

I have closed off my remaining issues with the documentation and installations - it all works well now. I went through the checklist and modified accordingly. Later this week I will check the visualisation to confirm the functionality and performance, but all looks good so far.
