+++ README.md
@@ -23,5 +23,6 @@ See current version and history [here](https://github.com/BenMql/coral/wiki/vers
 ## Benchmarks
 + [Rayleigh-Benard](https://github.com/BenMql/coral/wiki/benchmark01_rayleigh_benard)
 + [Rotating Rayleigh-Benard](https://github.com/BenMql/coral/wiki/benchmark02_RRBC_julien_JFM96)
++ [Rotating convective dynamo](https://github.com/BenMql/coral/wiki/benchmark03_rotating_convective_MHD)
 
 ![coral.jpg](http://www.normalesup.org/~benmiquel/misc/coral.jpg)
