+++ paper/paper.bib
@@ -66,6 +66,20 @@ pages={R1}}
   doi = {10.1103/PhysRevFluids.5.113702}
 }
 
+@article{stellmachPRE04,
+  title = {Cartesian convection driven dynamos at low Ekman number},
+  author = {Stellmach, Stephan and Hansen, Ulrich},
+  journal = {Phys. Rev. E},
+  volume = {70},
+  issue = {5},
+  pages = {056312},
+  numpages = {16},
+  year = {2004},
+  month = {Nov},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevE.70.056312}
+}
+
 @article{dedalus,
   title = {Dedalus: A flexible framework for numerical simulations with spectral methods},
   author = {Burns, Keaton J. and Vasil, Geoffrey M. and Oishi, Jeffrey S. and Lecoanet, Daniel and Brown, Benjamin P.},
