### 2021-03-17 #2978 @BenMql
Thank you all for your time and your valuable suggestions, which have spurred improvements in both the code, the doc, and the paper.
If you have more remarks, I am looking forward to hearing them!
