### 2021-03-20 BenMql/coral#11 @rhaas80
missing file in 4_data_visualisation after First Run
In 4_data_visualisation one is asked to visualize xy averaged data like so:

```
python plot_xyAvg_timeseries.py tFull 23 tFull 35
```

yet running command yields

```
plane_layer/test1$ python3 plot_xyAvg_timeseries.py tFull 23 tFull 35
Traceback (most recent call last):
  File "coral/data/plane_layer/test1/plot_xyAvg_timeseries.py", line 8, in <module>
    avgQty = np.fromfile('./Timeseries/'+argv[iarg+1] +'.bin', dtype=np.float_)
FileNotFoundError: [Errno 2] No such file or directory: './Timeseries/tFull.bin'
```

and the Timeseries directory contains only

```
plane_layer/test1$ ls Timeseries/
tFull_XYavg_z00023.dat  time.dat       uu_volAvg.dat  ww_volAvg.dat
tFull_XYavg_z00035.dat  tw_volAvg.dat  vv_volAvg.dat
```

for me.

This

```diff
diff --git a/etc/python_scripts/plot_xyAvg_timeseries.py b/etc/python_scripts/plot_xyAvg_timeseries.py
index 49f13b6..ab2fca5 100644
--- a/etc/python_scripts/plot_xyAvg_timeseries.py
+++ b/etc/python_scripts/plot_xyAvg_timeseries.py
@@ -4,8 +4,8 @@ from sys import argv

 plt.figure()
 time = np.fromfile('./Timeseries/time.dat', dtype=np.float_)
-for iarg in range(len(argv)-1):
-    avgQty = np.fromfile('./Timeseries/'+argv[iarg+1] +'.bin', dtype=np.float_)
+for iarg in range(0,len(argv)-1,2):
+    avgQty = np.fromfile('./Timeseries/'+argv[iarg+1]+'_XYavg_'+'z%05d'%int(argv[iarg+2])+'.dat', dtype=np.float_)
     plt.plot(time, avgQty [: time.shape[0]])
     del avgQty

```

makes the plot work
