### 2021-05-18 86f9f98
issue #11 --- fixed plotting routine
