+++ etc/python_scripts/plot_xyAvg_timeseries.py
@@ -4,8 +4,8 @@ from sys import argv
 
 plt.figure()
 time = np.fromfile('./Timeseries/time.dat', dtype=np.float_)
-for iarg in range(len(argv)-1):
-    avgQty = np.fromfile('./Timeseries/'+argv[iarg+1] +'.bin', dtype=np.float_)
+for iarg in range(0,len(argv)-1,2):
+    avgQty = np.fromfile('./Timeseries/'+argv[iarg+1]+'_XYavg_'+str(int(argv[iarg+2])).zfill(5)+'.dat', dtype=np.float_)
     plt.plot(time, avgQty [: time.shape[0]])
     del avgQty
 
