### 2021-05-18 acd0f05
issue #11 --- fixed plotting routine (missing z)
