### 2021-05-18 fda3aa2
issue #12 --- fixed dealiased routine (now Python version agnostic)
