### 2021-05-25 a4640b6
adding tests for transforms (use --checks at runtime)
