+++ src/layer_pencils/layer_pencils_main.f90
@@ -119,11 +119,28 @@ Program layer_pencils_main
    if (verbose) print*, '=========================='
    if (verbose) print*, '... Now performing some checks'
    if (verbose) print*, '=========================='
+   if (verbose) print*, '... '                                                                   
    if (verbose) print*, '... First checking the transforms'
-   if (verbose) print*, '... Method: initialise physical fields. Transform to spectral and back'
-   if (verbose) print*, '... Compute relative error between moments (first, second, and third) of'
-   if (verbose) print*, '... physical array before and after the back-and-forth transforms.'      
-   if (verbose) print*, '... Array below: Process rank, 1st, 2nd, and 3rd order moment errors'    
+   if (verbose) print*, '... Method: initialise one physical fields F. Transform to spectral and back'
+   if (verbose) print*, '...         G = iDFT( DFT( F )) '
+   if (verbose) print*, '... For each process, we compute the error of the first three moments '
+   if (verbose) print*, '... of G and F normalized by the moments of F. '
+   if (verbose) print*, '... More explicitly, for n=1,2,3, this quantity is:'
+   if (verbose) print*, '... '                                                                   
+   if (verbose) print*, '=========================='
+   if (verbose) print*, '          NX,NY,NZ                   NX,NY,NZ              '
+   if (verbose) print*, '            ===                        ===                 '
+   if (verbose) print*, '            \                 n        \                 n '
+   if (verbose) print*, '            /   [F(iz,iy,ix)]    __    /   [G(iz,iy,ix)]   '
+   if (verbose) print*, '            ===                        ===                 '
+   if (verbose) print*, '          ix,iy,iz                   ix,iy,iz              '
+   if (verbose) print*, ' E  = ______________________________________________________'
+   if (verbose) print*, '  n                NX,NY,NZ '
+   if (verbose) print*, '                     ===     '
+   if (verbose) print*, '                     \                 n '
+   if (verbose) print*, '                     /     F(iz,iy,ix)   '
+   if (verbose) print*, '                     ===                             '
+   if (verbose) print*, '                   ix,iy,iz '
    if (verbose) print*, '=========================='
    
    do i3 = lbound(main%linear_variables(1)%phys, dim=3),&
@@ -151,8 +168,9 @@ Program layer_pencils_main
    !                 third_moment,&
    !                 sum((main%linear_variables(1)%phys)**3),&
    !                 abs(third_moment- sum((main%linear_variables(1)%phys)**2))
-
-   print*, my_rank,& 
+   
+   4262 Format ('core ',I4,' | E1=',ES10.3,' | E2=',ES10.3,' | E3=',ES10.3)
+   write(*,4262)  my_rank,& 
                     abs((first_moment- sum(main%linear_variables(1)%phys))/first_moment),&
                     abs((second_moment- sum((main%linear_variables(1)%phys)**2))/second_moment),&
                     abs((third_moment- sum((main%linear_variables(1)%phys)**3))/third_moment)
