+++ etc/patch/patch_BM.sh
@@ -1,82 +1,0 @@
-#!/bin/bash
-sed -i '45i \ \ integer, parameter, public :: IN_CORE_CHEBYSHEV = 4 !BM_patch' src/fft_fftw3_f03.f90
-sed -i '148s/.*/\ \ \ else if \(\(format==PHYSICAL_IN_Z\).or.\(format==IN_CORE_CHEBYSHEV\)\) then !BM patch/' src/fft_fftw3_f03.f90
-sed -i '165s/.*/\ \ \ else if \(\(format==PHYSICAL_IN_Z\).or.\(format==IN_CORE_CHEBYSHEV\)\) then !BM patch/' src/fft_fftw3_f03.f90
-sed -i '215s/.*/\ \ \ else if \(\(format==PHYSICAL_IN_Z\).or.\(format==IN_CORE_CHEBYSHEV\)\) then !BM patch/' src/fft_fftw3_f03.f90
-sed -i '523i \ \ \ \ else if \(format == IN_CORE_CHEBYSHEV\) then           ! BM patch' src/fft_fftw3_f03.f90
-sed -i '524i \ \ \ \ \ \ \ ! For R2C/C2R tranforms                           ! BM patch' src/fft_fftw3_f03.f90
-sed -i '525i \ \ \ \ \ \ \ call r2c_1m_z_plan\(plan\(0,3\), ph, sp\)             ! BM patch' src/fft_fftw3_f03.f90
-sed -i '526i \ \ \ \ \ \ \ call c2c_1m_y_plan\(plan\(0,2\), sp, FFTW_FORWARD \)  ! BM patch' src/fft_fftw3_f03.f90
-sed -i '527i \ \ \ \ \ \ \ call c2c_1m_y_plan\(plan\(2,2\), sp, FFTW_BACKWARD\)  ! BM patch' src/fft_fftw3_f03.f90
-sed -i '528i \ \ \ \ \ \ \ call c2r_1m_z_plan\(plan\(2,3\), sp, ph\)             ! BM patch' src/fft_fftw3_f03.f90
-sed -i '852i \ \ \ \ \ \ \    ! ############################################### ! BM patch' src/fft_fftw3_f03.f90
-sed -i '853i \ \ \ \ \ \ \    else if \(format==IN_CORE_CHEBYSHEV\) then          ! BM patch' src/fft_fftw3_f03.f90
-sed -i '854i \ \ \ \ \ \ \                                                      ! BM patch' src/fft_fftw3_f03.f90
-sed -i '855i \ \ \ \ \ \ \       ! ===== 1D FFTs in Z =====                     ! BM patch' src/fft_fftw3_f03.f90
-sed -i '856i \ \ \ \ \ \ \       call r2c_1m_z\(in_r,wk13\)                       ! BM patch' src/fft_fftw3_f03.f90
-sed -i '857i \ \ \ \ \ \ \                                                      ! BM patch' src/fft_fftw3_f03.f90
-sed -i '858i \ \ \ \ \ \ \       ! ===== Swap Z --> Y; 1D FFTs in Y =====       ! BM patch' src/fft_fftw3_f03.f90
-sed -i '859i \ \ \ \ \ \ \       if \(dims\(1\)>1\) then                            ! BM patch' src/fft_fftw3_f03.f90
-sed -i '860i \ \ \ \ \ \ \          call transpose_z_to_y\(wk13,wk2_r2c,sp\)      ! BM patch' src/fft_fftw3_f03.f90
-sed -i '861i \ \ \ \ \ \ \          call c2c_1m_y\(wk2_r2c,-1,plan\(0,2\)\)         ! BM patch' src/fft_fftw3_f03.f90
[…]
