### 2021-06-10 e27f0d2
issue  #12 --- fixing deprecated matplotlib features in data viz. routines
