### 2021-06-10 #2978 @BenMql
Dear all,

The issue that have been thoroughly reported by @rhaas80  and @robertsawko  are now fixed.

IMHO, only issue #13 still deserves some additional work: I tried my best to have a pedagogical and gradual approach for teaching the user how to enter PDEs in the solver Coral. I am still working to streamline the process. On the bright side, the user is now provided with 3 full benchmarks taken from some (recent) published references, which provide some starting points for implementing their own equations (with the help of the doc).

On a different note, while testing on different machines since my last update, I triggered a bug of the DFT library with some recent compilers. After a while, it became clear that the culprits were optimization flags, which should be rolled back a little with recent gcc compilers when compiling `2decomp&fft`. This is now properly documented in the wiki. Perhaps more interestingly, this has prompted me to implement automated tests in Coral, that allows the user to test their compilation of the DFT library. @eloisabentivegna Apologies for the additional delay, this buggy behaviour was completely unexpected and was completely under the radar until very recently.
