### 2021-06-10 #2978 @eloisabentivegna
Thanks for the update, @BenMql! I am glad this process has prompted you to further develop Coral (for one reason or another!).

@rhaas80, @robertsawko, do you have any comments?
