### 2021-06-15 787cd26
mean flows can be renormalized; fixed U(rms) on display; removed unused .o file from Makefile
