### 2021-06-16 d0427bf
reduced amplitude of the (noisy) initial condition, for easier transients
