### 2021-06-17 79fbb78
adding runtime flags (not fully tested yet)
