### 2021-06-23 9c93888
fixed bug for entering NL term for zero-mode eqn
