### 2021-07-01 db8db1e
linear_variable_full can involve derivatives of order > 1
