### 2021-07-02 23e9c9d
x- and y-differentiation methods fixed in process_data.py
