### 2021-07-05 97e7381
minor tweak, plotting routines
