### 2021-07-07 3180e97
fixed minor bugs related to sources; implemented additional check for proper equation entries
