### 2021-07-25 #2978 @robertsawko
Hello,

Apologies again for delay. I have now finished reviewing the code. Few minor points to consider below. Overall I think the code is an excellent contribution.

 * As far as I can see there are no automated tests for the code. I would suggest @BenMql considers whether any low-level tests can be proposed for some aspects of the solver and try some standard testing suites. If that is not possible, then I would suggest putting example outputs of the visualised quantities in [here](https://github.com/BenMql/coral/wiki/4_data_visualisation).
 *  I could not verify the performance claims as I do not have access to large enough machine currently, but @BenMql provides a reference to another publication where they document a performance study. I find this entirely sufficient and therefore I am considering this task completed.
 *  The table of content of the wiki pages is becoming a little long. I would consider simplifying it with a custom sidebar as described in [here](https://docs.github.com/en/communities/documenting-your-project-with-wikis/creating-a-footer-or-sidebar-for-your-wiki#creating-a-sidebar).

I believe several outstanding issues can be also closed now. For instance [this](https://github.com/BenMql/coral/issues/11) is working for me on the latest commit.
