### 2021-07-28 #2978 @danielskatz
@rhaas80 - when do you think you will be able to make progress on the remaining items in your review?
