### 2021-08-11 BenMql/coral#20 @rhaas80
no quantitative way to test functionality of the software
The first run and benchmark examples provide a way to test that one's installation of Coral works as expected, yet they only provide graphs requiring that one eyeball compares one's results with the image files in the documentation eg https://github.com/BenMql/coral/wiki/benchmark01_rayleigh_benard

For https://github.com/openjournals/joss-reviews/issues/2978 it would be good if some of the data files of the benchmark runs were also provided so that a more quantitative comparison (even if only plotting both datasets on the same graph) was possible.
