### 2021-08-13 BenMql/coral#21 @rhaas80
no community guideline
I seem to be unable to find guidelines for potential contributors. Both technical (make a pull request, create a ticket with a patch, quality expectations, any required formatting style of code) and non-technical (is any transfer of copyright expected?).

related to: https://github.com/openjournals/joss-reviews/issues/2978
