### 2021-08-23 #2978 @danielskatz
👋 Hi all - Can I get an update on the status of this review?

@rhaas80  - it looks like you opened some issues 2 weeks ago. Once these are resolved, will your review be complete?  Or are there other things you are also doing?

@robertsawko - What is needed for you to complete your review?  Are the items in https://github.com/openjournals/joss-reviews/issues/2978#issuecomment-886235689 from about a month ago still outstanding?

@BenMql - What is your progress on the issues raised by @rhaas80 and those in the comment from @robertsawko?

Thanks all!!
