### 2021-08-27 #2978 @rhaas80
I just finished reading through the source. Looks good to me. Overall quite readable. For some files (most of the LP Coral file) there are haders:
```
 !=============================================================================
 !                            C O R A L
 !=============================================================================
 !
 ! MODULE: LP_IMEX_timestepping
 !
 !> @author
 !> Ben Miquel, www.normalesup.org/~benmiquel, benjamin.miquel@tutanota.com
 !
 ! DESCRIPTION
 !> core routines and data structure implementing the time-marching schemes
 !> of Ascher-Ruuth-Spiteri 1997 (Applied Numerical Mathematics 25, pp151-167)
 !> [denoted below as ARS97].
 !
 !=============================================================================
```
but not quite for all. Since these are very useful to get a rough idea what may be expected in a file, having them for all files would be great (but purely optional).
