### 2021-08-30 b10c470
Create CONTRIBUTING.md
