### 2021-08-30 BenMql/coral#21 @BenMql
@rhaas80 , good point. I have added a `CONTRIBUTING.md` file. With time and as the number of contributors/users increases, the guidelines will evolve.
I leave this issue open for a week, in case you find something missing.
