### 2021-08-30 BenMql/coral#20 @BenMql
All the ways available for the users to test their installation are now gathered [on this page](https://github.com/BenMql/coral/wiki/validation).

I thought I had given all the numerical values, but it turns out that some were missing, indeed. This is now fixed so that validations should now be straightforward for users.

Although not entirely automated, these tests require very few lines of codes. The pattern is often: `cp` the content of a repository, `mpiexec` the executable, `python` a provided routine. It is then possible to compare the numbers displayed to the reference value provided in the wiki.

In particular, the graphs displayed in the wiki of benchmark1 can be generated by the user, with the analytic prediction superimposed. The initial condition has been changed since a few weeks so that the initial exponential growth is even longer and matches well the analytic prediction (on a test that I just run).
