### 2021-08-30 #2978 @BenMql
:wave: all,

I believe I have addressed all the remaining issues, which I summarize here:

+ both @rhaas80 and @robertsawko mentioned the lack of quantitative tests (e.g. [here](https://github.com/BenMql/coral/issues/20)). I have now gathered all the tests in a dedicated wiki page [here](./validation). For each benchmark, users are provided with a python routine so that the computation of global quantities is straightforward, and these values can readily be compared to reference values given in the same [wiki page](./validation).
+ both @rhaas80 and @robertsawko mentioned the lack of community guidelines (e.g. [here](https://github.com/BenMql/coral/issues/21). A `CONTRIBUTING.md` file has been added to fill this void and address this issue.
+ The third item missing in @robertsawko 's checklist is:  "Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)" This is a bit of a grey area, but perhaps this item could be checked off as there is indeed no claim for Coral itself (although `2decomp&fft` has been benchmarked by their authors). In the long term, a systematic performance study will be available on the wiki. For now, I do not have enough data points though.
+ @robertsawko mentioned the relevance of a custom sidebar (as the number of pages increases), which I have happily implemented.
+ @rhaas80 mentioned the lack of consistency regarding headers in the source file, as an optional comment. This is a good point. More generally, these headers are part of the documentation of the code written in Doxygen. This documentation is still work in progress but will hopefully cover all source files and subroutines (at which point the headers will be homogeneous).

Apologies for this long interruption during Summer. Again, I want to express my gratitude for your suggestions which are very helpful and result in improved sources/doc.
Let me know if I have missed something.
