### 2021-08-30 #2978 @rhaas80
contribution.md looks very good. While no automated tests are provided the manual tests now come with a textual description of the expected numerical output that lets users manually verify that their copy of CORAL works as expected.

I have no further comments that would delay publication.
