### 2021-09-02 #2978 @robertsawko
Hello and thanks to @danielskatz for a ping on another email.

@BenMql thanks for addressing all of the comments. I think the wiki looks very clear now and examples are well described. I don't have any further comments. I've checked the remaining boxes.
