### 2021-09-02 #2978 @danielskatz
@BenMql - I'll proofread the paper shortly, then give you a few final steps (perhaps including suggested text changes) to do.
