+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{miquelJFM20, 
-title={On the role of the Prandtl number in convection driven by heat sources and sinks}, 
+title={On the role of the {P}randtl number in convection driven by heat sources and sinks}, 
 volume={900}, 
 DOI={10.1017/jfm.2020.485}, 
 journal={J. Fluid Mech.}, 
@@ -42,7 +42,7 @@ pages={R1}}
   }
 
 @article{julienJFM96,
-  title={Rapidly rotating turbulent Rayleigh-B{\'e}nard convection},
+  title={Rapidly rotating turbulent {R}ayleigh-{B}{\'e}nard convection},
   author={Julien, K. and Legg, S. and McWilliams, J. and Werne, J.},
   journal={J. Fluid Mech.},
   volume={322},
@@ -67,7 +67,7 @@ pages={R1}}
 }
 
 @article{stellmachPRE04,
-  title = {Cartesian convection driven dynamos at low Ekman number},
+  title = {Cartesian convection driven dynamos at low {E}kman number},
   author = {Stellmach, Stephan and Hansen, Ulrich},
   journal = {Phys. Rev. E},
   volume = {70},
@@ -118,7 +118,7 @@ url={http://nek5000.mcs.anl.gov}}
   number = {31},
   pages = {1071},
   author = {Mikael Mortensen},
-  title = {Shenfun: High performance spectral Galerkin computing platform},
+  title = {Shenfun: High performance spectral {G}alerkin computing platform},
   journal = {Journal of Open Source Software}
 }
 
@@ -147,7 +147,7 @@ year = {2019},
 publisher = {Ubiquity Press,  Ltd.},
 volume = {7},
 author = {Pierre Augier and Ashwin Vishnu Mohanan and Cyrille Bonamy},
-title = {{FluidDyn}: A Python Open-Source Framework for Research and Teaching in Fluid Dynamics
+title = {{FluidDyn}: A {P}ython Open-Source Framework for Research and Teaching in Fluid Dynamics
     by Simulations,  Experiments and Data Processing},
 journal = {Journal of Open Research Software}
 }
@@ -160,8 +160,8 @@ publisher = {Ubiquity Press,  Ltd.},
 volume = {7},
 author = {Mohanan, Ashwin Vishnu and Bonamy, Cyrille and Linares, Miguel
     Calpe and Augier, Pierre},
-title = {{FluidSim}: {Modular}, {Object}-{Oriented} {Python} {Package} for
-    {High}-{Performance} {CFD} {Simulations}},
+title = {{FluidSim}: Modular, Object-Oriented {P}ython Package for
+    High-Performance {CFD} Simulations},
 journal = {Journal of Open Research Software}
 }
 
