### 2021-09-02 #2978 @danielskatz
@BenMql  - I've suggested some changes in https://github.com/BenMql/coral/pull/22 - please either merge these or let me know what you disagree with
