### 2021-09-03 #2978 @danielskatz
@BenMql - I missed one issue, which will be fixed if you merge https://github.com/BenMql/coral/pull/23 (I hope) - Please merge this, and run `@whedon generate pdf` to check if it looks ok.  If so, then please go ahead and
