### 2021-09-03 #2978 @danielskatz
@BenMql - you might want to change the bib entry for your J Fluid Mech paper to have your first initial rather than first name for consistency, but it's not a big deal either way
