+++ paper/paper.bib
@@ -5,7 +5,7 @@ volume={900},
 DOI={10.1017/jfm.2020.485}, 
 journal={J. Fluid Mech.}, 
 publisher={Cambridge University Press}, 
-author={Miquel, Benjamin and Bouillaut, Vincent and Aumaître, Sébastien and Gallet, Basile}, 
+author={Miquel, B. and Bouillaut, V. and Aumaître, S. and Gallet, B.}, 
 year={2020}, 
 pages={R1}}
 
