### 2021-09-03 #2978 @BenMql
@danielskatz That looks better. The version tag is v1.1.12
What do people customarily do: register the upload on Zenodo as software or publication? (Apologies for the stupid question, but after all there is a paper and a code in the archive.)
