### 2021-09-05 #2978 @BenMql
@danielskatz -- the archived Coral v1.1.12 has been uploaded on Zenodo and has received the DOI `10.5281/zenodo.5458888`
